# [multibody] Simplify frames in MultibodyPlant