--- a/multibody/tree/frame_body_pose_cache.cc
+++ b/multibody/tree/frame_body_pose_cache.cc
@@ -1,0 +1,4 @@
+#include "drake/multibody/tree/frame_body_pose_cache.h"
+
+DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
+    class ::drake::multibody::internal::FrameBodyPoseCache);