This is a small step towards a high-performance multibody kernel, see issue #18442. An important goal is to eliminate Context spelunking and virtual methods from the multibody kernel; this moves us part way there and provides a modest performance boost (benchmarks below).

### Background
The Frame class hierarchy was originally designed to support frames on deformable bodies, supporting the notion that a frame's pose on its body might not be constant. We don't need that abstraction now -- all our Frames have fixed poses after parameters have been set (they are either RigidBodyFrames or FixedOffsetFrames). There is overhead associated with that abstraction -- virtual methods, runtime recalculations of pose, and copies rather than references.

### What's here
This PR removes the deformable-frame abstraction and then takes advantage of that to cache Frame pose information as soon as parameters are set rather than at run time. Specifically:
- Removes useless FrameBase class.
- Caches frame pose-in-body and inverse.
- Replaces runtime virtuals and computations with inline cache references.
- Now no kernel function looks in the Context for Frame body poses.

### What's not here
- There are still many places where a Context is being passed in to kernel functions. All of those need to be replaced with precalculated cache entries so that the kernel operates only on numerical quantities. TODOs for now.
- BodyNode needs to be templatized by concrete Mobilizer to allow fixed-size vectors and inline calls rather than virtual functions.

### Release note
Although FrameBase was a public class it had no functionality and served only as the useless base class for Frame. There are no direct uses of FrameBase in Drake or Anzu, and no Python bindings. So although this is technically a breaking change I'm not deprecating it because there is ~zero chance anyone will be affected.

### Benchmarking
CassieBench measures only multibody kernel time, in microseconds.

|      Test        | Master/GCC  |  This PR | Speedup |  Master/Clang  | ThisPR | Speedup |
| --------------- | ------------------ | ----------- | ------------ | -------------------- | ---------- | ------------ |
| MassMatrix |        15.7        |    13.6     |    13%     |         11.5          |   10.4    |    10%    |
| InverseDynamics |        18.3        |    15.3     |    16%     |         13         |   11    |    15%    |
| ForwardDynamics |        40.8        |    38.1     |    7%     |         37.1          |   35.5    |    4%    |

AutoDiff and Expression benchmarks were not measurably affected by this change.

- Puget Xeon E5-2699 v4 @2.2GHz
- gcc 11.4.0
- clang 15.0.7

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RobotLocomotion/drake/21853)
<!-- Reviewable:end -->